gcc: remove more irrelevant instructions. add jmps back
The `jump` pattern was accidently removed in the last patch. Also, the
`indirect_jump` pattern is a required pattern, so I added it back, but
removed the speculation store barrier. Miscellaneous other instructions
were removed. Also, we need to remove the corresponding functions in
`aarch64.cc` that reference the ones generated by the machine
description.